Cleaning up my queries to use pool.query and abstract out pool.connect and release. Move s3 photo storage to a helper function.
