+++ api/helpers.js
@@ -1,4 +1,5 @@
 const multer = require('multer');
+const Aws = require('aws-sdk');
 
 module.exports = {
 
@@ -8,5 +9,28 @@ module.exports = {
     } else {
       cb(null, false);
     }
+  },
+
+  storePhoto: (req) => {
+    const s3 = new Aws.S3({
+      accessKeyId: process.env.AWSKEY,
+      secretAccessKey: process.env.AWSSECRET
+    });
+    const params = {
+      Bucket: process.env.AWSBUCKET,
+      Key: req.file.originalname,
+      Body: req.file.buffer,
+      ACL: 'public-read-write',
+      ContentType: 'image/jpeg'
+    };
+    return new Promise((resolve, reject) => {
+      s3.upload(params, (err, result) => {
+        if (err) {
+          console.log(err);
+          reject(null);
+        }
+        resolve(result);
+      });
+    });
   }
 };
